requirements.txt
gradio>=4.0.0
pandas>=1.5.0
numpy>=1.21.0
faiss-cpu>=1.7.0
requests>=2.28.0
openai>=1.0.0
groq>=0.4.0
nbformat>=5.0.0